% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sentiment.R
\name{calculate_sentiment}
\alias{calculate_sentiment}
\title{Calculates the sentiment values for each sentences on by one}
\usage{
calculate_sentiment(sentences)
}
\arguments{
\item{sentences}{vector of strings for which the sentiment values
should be calculated}
}
\value{
vector of sentiment values corresponding to the inout
sentences
}
\description{
Calculates the sentiment values for each sentences on by one
}
