% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_manipulation.R
\name{calculate_word_frequencies}
\alias{calculate_word_frequencies}
\title{Calculates the global frequencies of words in the given sentences}
\usage{
calculate_word_frequencies(sentences)
}
\arguments{
\item{sentences}{vector of strings for which the word frequencies
should be calculated}
}
\value{
data.frame with word frequencies
}
\description{
Calculates the global frequencies of words in the given sentences
}
