% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_manipulation.R
\name{clean_sentences}
\alias{clean_sentences}
\title{Cleans input senteces from html, converts to lower
characters and simplifies white spaces.}
\usage{
clean_sentences(sentences)
}
\arguments{
\item{sentences}{vector of input sentences}
}
\value{
vector of words
}
\description{
Cleans input senteces from html, converts to lower
characters and simplifies white spaces.
}
