% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_comments_graph_plot.R
\name{mod_comments_graph_plot_ui}
\alias{mod_comments_graph_plot_ui}
\alias{mod_comments_graph_plot_server}
\title{mod_comments_graph_plot_ui and mod_comments_graph_plot_server}
\usage{
mod_comments_graph_plot_ui(id, title)

mod_comments_graph_plot_server(input, output, session, comments_promise,
  selected_comment)
}
\arguments{
\item{id}{shiny id}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}
}
\description{
A shiny Module.
}
\keyword{internal}
