% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_comments_table.R
\name{mod_comments_table_ui}
\alias{mod_comments_table_ui}
\alias{mod_comments_table_server}
\title{mod_comments_table_ui and mod_comments_table_server}
\usage{
mod_comments_table_ui(id)

mod_comments_table_server(input, output, session, comments_promise)
}
\arguments{
\item{id}{shiny id}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}
}
\description{
A shiny Module.
}
\keyword{internal}
