% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_stories_table.R
\name{mod_stories_table_ui}
\alias{mod_stories_table_ui}
\alias{mod_stories_table_server}
\title{mod_stories_table_ui and mod_stories_table_server}
\usage{
mod_stories_table_ui(id)

mod_stories_table_server(input, output, session)
}
\arguments{
\item{id}{shiny id}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}
}
\description{
A shiny Module.
}
\keyword{internal}
