% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_wordcloud_plot.R
\name{mod_wordcloud_plot_ui}
\alias{mod_wordcloud_plot_ui}
\alias{mod_wordcloud_plot_server}
\title{mod_wordcloud_plot_ui and mod_wordcloud_plot_server}
\usage{
mod_wordcloud_plot_ui(id, title)

mod_wordcloud_plot_server(input, output, session, comments_promise)
}
\arguments{
\item{id}{shiny id}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}
}
\description{
A shiny Module.
}
\keyword{internal}
