% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_manipulation.R
\name{remove_html}
\alias{remove_html}
\title{Removes html elements from the input string}
\usage{
remove_html(texts)
}
\arguments{
\item{texts}{vector of input string to remove html from}
}
\value{
vector of strings with removed html elements
}
\description{
Removes html elements from the input string
}
