% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_manipulation.R
\name{remove_stop_words}
\alias{remove_stop_words}
\title{Removes stop words from the given vector of words}
\usage{
remove_stop_words(words)
}
\arguments{
\item{words}{vector of words from which stop words
should be removed}
}
\value{
vector of words with stop_words removed
}
\description{
Removes stop words from the given vector of words
}
